refactor(runtime): move runtime runners out of presentation
