+++ src/presentation/ui/tui/internal/bubble_tea/header_label_test.go
@@ -2,7 +2,7 @@ package bubble_tea
 
 import (
 	"testing"
-	"tungo/presentation/runners/version"
+	"tungo/runtime/version"
 )
 
 func TestProductLabel_DefaultWhenVersionUnset(t *testing.T) {
